--- src/components/AccountChoiceModular/index.module.scss
@@ -1,77 +1,0 @@
-.AccountChoiceModularWrapper {
-  margin: 30px auto;
-  position: relative;
-  max-width: 600px;
-  padding: 30px 60px;
-  border: 1px solid #2596be;
-  border-radius: 20px;
-  >a :first-child{
-    position: absolute;
-    right: 20px;
-    top: 20px;
-    font-size: 22px;
-    cursor: pointer;
-    color: black;
-    &:hover{
-      scale: .85;
-    }
-  }
-  > h2,
-  > p {
-    text-align: center;
-  }
-
-  > p {
-    font-size: 14px;
-    a {
-      color: #2596be;
-    }
-  }
-  * {
-    outline-color: #2596be;
-  }
-  .AccountChoiceModular {
-    display: flex;
-    flex-direction: column;
-    gap: 20px;
-    margin-block: 25px;
-    > div {
-      border: 1px solid rgba(0, 0, 0, 0.2);
-      padding: 0px 10px;
-      border-radius: 5px;
-      display: flex;
-      justify-content: space-between;
-      width: 100%;
-      height: 50px;
-      label {
-        display: flex;
-        align-items: center;
-        font-size: 16px;
-        width: 98.5%;
-        height: 100%;
-        cursor: pointer;
-      }
-      input {
-        height: 16px;
-        width: auto;
-      }
-    }
-  }
-  .ContinueButton {
-    width: 80%;
-    max-width: 500px;
-    border: none;
-    padding: 15px 12px;
-    border-radius: 4px;
-    color: white;
-    font-size: 12px;
-    font-weight: 500;
-    margin: auto;
-    margin-block: 10px;
-    margin-top: 30px;
-    text-transform: capitalize;
-    display: block;
-    text-align: center;
-    border-radius: 5px;
-  }
-}
