+++ src/components/Notice/index.module.scss
@@ -1,0 +1,5 @@
+.Notice{
+  background-color: White;
+  color: darkgray;
+  font-size: 14px;
+}
