+++ src/components/SelectAccountType/index.jsx
@@ -1,0 +1,101 @@
+import s from "./index.module.scss";
+import { useState } from "react";
+import { Link, useNavigate } from "react-router-dom";
+import RadioButton from "../RadioButton";
+import { AiOutlineClose } from "react-icons/ai";
+import Logo from "../../static/images/logo_colored.png";
+import axios from "axios";
+const SelectAccountType = () => {
+  const navigate = useNavigate();
+  const { email, authKey } = JSON.parse(
+    window.localStorage.getItem("NXGJOBHUBLOGINKEYV1")
+  );
+  const [accountChoice, setAccountChoice] = useState("");
+  const accountTypes = {
+    techtalent:
+      "https://job-hub-591ace1cfc95.herokuapp.com/api/v1/tech-talent/register/",
+    agent: "https://job-hub-591ace1cfc95.herokuapp.com/api/agents/createAgent",
+    employer: "https://job-hub-591ace1cfc95.herokuapp.com/api/v1/auth/get-user",
+  };
+  const handleChange = (e) => {
+    const { value } = e.target;
+    setAccountChoice(value);
+  };
+  const setAccountType = () => {
+    axios
+      .post(
+        accountTypes[accountChoice],
+        {
+          email,
+        },
+        {
+          headers: {
+            authorization: authKey,
+          },
+        }
+      )
+      .then(navigate("/dashboard")).catch(err => console.log(err));
+  };
+  return (
+    <>
+      <div className={s.page}>
+        <nav className={s.navBar}>
+          <img src={Logo} alt="" />
+        </nav>
+        <div className={s.main}>
+          <div className={s.AccountChoiceModularWrapper}>
+            <h2>Join our community of professionals</h2>
+            <p>Select an account type that suits you.</p>
+
+            <Link to="/login" title="Close">
+              {" "}
+              <AiOutlineClose />{" "}
+            </Link>
+            <div className={s.AccountChoiceModular}>
+              <RadioButton
+                label={"I'm a tech talent"}
+                value={"techtalent"}
+                id={"techtalent"}
+                name={"account_choice"}
+                onchange={handleChange}
+              />
+              <RadioButton
+                label={"I'm an employer"}
+                value={"employer"}
+                id={"employer"}
+                name={"account_choice"}
+                onchange={handleChange}
+              />
+              <RadioButton
+                label={"I'm an agent"}
+                value={"agent"}
+                id={"agent"}
+                name={"account_choice"}
+                onchange={handleChange}
+              />
+            </div>
+            <button
+              className={s.ContinueButton}
+              onClick={setAccountType}
+              style={{
+                backgroundColor: accountChoice
+                  ? "#2596be"
+                  : "rgb(194, 192, 192)",
+              }}
+              aria-disabled={accountChoice ? false : true}
+              disabled={accountChoice ? false : true}
+              to={`./${accountChoice}`}
+            >
+              Continue
+            </button>
+            <p>
+              Already have an account? <Link to={"/login"}> Log in </Link>
+            </p>
+          </div>
+        </div>
+      </div>
+    </>
+  );
+};
+
+export default SelectAccountType;
