+++ src/components/SelectAccountType/index.module.scss
@@ -1,0 +1,98 @@
+
+.navBar {
+  display: flex;
+  width: 100%;
+  margin:auto;
+  background-color: #2596be;
+  padding: 10px 25px;
+  img {
+    height: 65px;
+    width: auto;
+  }
+}
+
+.main {
+  margin: 0px auto 0px;
+  width: 95%;
+  background-color: #f7f5f55a;
+  border: 1.5px solid #ffffff;
+  min-height: 80vh;
+  .AccountChoiceModularWrapper {
+    margin: 30px auto;
+    position: relative;
+    max-width: 600px;
+    padding: 30px 60px;
+    border: 1px solid #2596be;
+    border-radius: 20px;
+    >a :first-child{
+      position: absolute;
+      right: 20px;
+      top: 20px;
+      font-size: 22px;
+      cursor: pointer;
+      color: black;
+      &:hover{
+        scale: .85;
+      }
+    }
+    > h2,
+    > p {
+      text-align: center;
+    }
+  
+    > p {
+      font-size: 14px;
+      button {
+        color: #2596be;
+        cursor: pointer;
+      }
+    }
+    * {
+      outline-color: #2596be;
+    }
+    .AccountChoiceModular {
+      display: flex;
+      flex-direction: column;
+      gap: 20px;
+      margin-block: 25px;
+      > div {
+        border: 1px solid rgba(0, 0, 0, 0.2);
+        padding: 0px 10px;
+        border-radius: 5px;
+        display: flex;
+        justify-content: space-between;
+        width: 100%;
+        height: 50px;
+        label {
+          display: flex;
+          align-items: center;
+          font-size: 16px;
+          width: 98.5%;
+          height: 100%;
+          cursor: pointer;
+        }
+        input {
+          height: 16px;
+          width: auto;
+        }
+      }
+    }
+    .ContinueButton {
+      width: 80%;
+      max-width: 500px;
+      border: none;
+      padding: 15px 12px;
+      border-radius: 4px;
+      color: white;
+      font-size: 12px;
+      font-weight: 500;
+      margin: auto;
+      margin-block: 10px;
+      margin-top: 30px;
+      text-transform: capitalize;
+      display: block;
+      text-align: center;
+      border-radius: 5px;
+    }
+  }
+}
