--- src/pages/CreateAccount/index.module.scss
@@ -1,19 +1,0 @@
-.navBar {
-  display: flex;
-  width: 100%;
-  margin:auto;
-  background-color: #2596be;
-  padding: 10px 25px;
-  img {
-    height: 65px;
-    width: auto;
-  }
-}
-
-.main {
-  margin: 0px auto 0px;
-  width: 95%;
-  background-color: #f7f5f55a;
-  border: 1.5px solid #ffffff;
-  min-height: 80vh;
-}
