+++ src/pages/Register/RegistrationForm.module.scss
