+++ src/pages/Register/index.module.scss
@@ -1,5 +1,5 @@
 .page {
-  background-image: url(../../../static/images/sign_up_bg.svg);
+  background-image: url(../../static/images/sign_up_bg.svg);
   background-color: black;
   background-repeat: no-repeat;
   background-size: cover;
